feat(event): retrieve and display lowest ticket price

+++ src/main/java/com/ramarizdev/eventureBackend/event/dto/EventSummaryDto.java
@@ -29,6 +29,6 @@ public class EventSummaryDto {
 
     private String category;
 
-    private Integer lowestPrice;
+    private Double lowestPrice;
 
 }
